Data.Row

Description

This module includes a set of common functions for Records and Variants. It includes:

Common constructors, destructors, and querying functions

It specifically excludes:

Functions that have the same name for Records and Variants (e.g. focus, update, fromLabels, etc.)

Common clashes with the standard Prelude or other modules (e.g. map, sequence, zip, Map, etc.)

If these particular functions are needed, they should be brought in qualified from one of the Data.Row.*** modules directly.

Types and constraints

data Label (s :: Symbol)

A label

Constructors

Label

class KnownSymbol (n :: Symbol)

This class gives the string associated with a type-level symbol. There are instances of the class for every concrete literal: "hello", etc.

Since: base-4.7.0.0

Minimal complete definition

symbolSing

type family AllUniqueLabels (r :: Row k) :: Constraint where ...

Are all of the labels in this Row unique?

Equations

AllUniqueLabels (R r) = AllUniqueLabelsR r

type WellBehaved ρ = (Forall ρ Unconstrained1, AllUniqueLabels ρ)

A convenient way to provide common, easy constraints

data Var (r :: Row *)

The variant type.

data Rec (r :: Row *)

A record with row r.

data Row a

The kind of rows. This type is only used as a datakind. A row is a typelevel entity telling us which symbols are associated with which types.

type Empty = R '[]

Type level version of empty

type (≈) a b = a ~ b infix 4

A lower fixity operator for type equality

class (r .! l) ≈ a => HasType l a r

Alias for (r .! l) ≈ a. It is a class rather than an alias, so that it can be partially applied.

type family Subset (r1 :: Row k) (r2 :: Row k) :: Constraint where ...

Is the first row a subset of the second? Or, does the second row contain every binding that the first one does?

Equations

Subset ('R '[]) r = Unconstrained

Subset ('R ((l ':-> a) ': x)) r = ((r .! l) ≈ a, Subset ('R x) r)

class Lacks (l :: Symbol) (r :: Row *)

Alias for .\. It is a class rather than an alias, so that it can be partially applied.

type family (r :: Row k) .\ (l :: Symbol) :: Constraint where ... infixl 4

Does the row lack (i.e. it does not have) the specified label?

Equations

(R '[]) .\ l = Unconstrained

(R r) .\ l = LacksR l r r

type family (l :: Row k) .+ (r :: Row k) :: Row k where ... infixl 6

Type level Row append

Equations

x .+ (R '[]) = x

(R '[]) .+ y = y

x .+ (R '[l :-> a]) = Extend l a x

(R '[l :-> a]) .+ y = Extend l a y

(R l) .+ (R r) = R (Merge l r)

type family (l :: Row k) .\/ (r :: Row k) where ... infixl 6

The minimum join of the two rows.

Equations

x .\/ (R '[]) = x

(R '[]) .\/ y = y

(R l) .\/ (R r) = R (MinJoinR l r)

type family (l :: Row k) .\\ (r :: Row k) :: Row k where ... infixl 6

Type level Row difference. That is, l .\\ r is the row remaining after removing any matching elements of r from l.

Equations

(R l) .\\ (R r) = R (Diff l r)

type family (l :: Row k) .// (r :: Row k) where ... infixl 6

The overwriting union, where the left row overwrites the types of the right row where the labels overlap.

Equations

x .// (R '[]) = x

(R '[]) .// y = y

(R l) .// (R r) = R (ConstUnionR l r)

class BiForall (r1 :: Row k1) (r2 :: Row k2) (c :: k1 -> k2 -> Constraint)

Any structure over two rows in which the elements of each row satisfy some constraints can be metamorphized into another structure over both of the rows.

Minimal complete definition

biMetamorph

class Forall (r :: Row k) (c :: k -> Constraint)

Any structure over a row in which every element is similarly constrained can be metamorphized into another structure over the same row.

Minimal complete definition

metamorph

type FreeForall r = Forall r Unconstrained1

FreeForall can be used when a Forall constraint is necessary but there is no particular constraint we care about.

type FreeBiForall r1 r2 = BiForall r1 r2 Unconstrained2

FreeForall can be used when a BiForall constraint is necessary but there is no particular constraint we care about.

switch :: forall v r x. BiForall r v (AppliesTo x) => Var v -> Rec r -> x

A Var and a Rec can combine if their rows line up properly. Given a Variant along with a Record of functions from each possible value of the variant to a single output type, apply the correct function to the value in the variant.

caseon :: forall v r x. BiForall r v (AppliesTo x) => Rec r -> Var v -> x

The same as switch but with the argument order reversed

Record Construction

empty :: Rec Empty

The empty record

type (.==) (l :: Symbol) (a :: k) = Extend l a Empty infix 7

A type level way to create a singleton Row.

(.==) :: KnownSymbol l => Label l -> a -> Rec (l .== a) infix 7

The singleton record

pattern (:==) :: forall l a. KnownSymbol l => Label l -> a -> Rec (l .== a) infix 7

A pattern for the singleton record; can be used to both destruct a record when in a pattern position or construct one in an expression position.

Restriction

type family (r :: Row k) .- (s :: Symbol) :: Row k where ... infixl 6

Type level Row element removal

Equations

(R r) .- l = R (Remove l r)

(.-) :: KnownSymbol l => Rec r -> Label l -> Rec (r .- l) infixl 6

Record restriction. Remove the label l from the record.

Query

type family (r :: Row k) .! (t :: Symbol) :: k where ... infixl 5

Type level label fetching

Equations

(R r) .! l = Get l r

(.!) :: KnownSymbol l => Rec r -> Label l -> r .! l

Record selection

Union

(.+) :: forall l r. FreeForall l => Rec l -> Rec r -> Rec (l .+ r) infixl 6

Record disjoint union (commutative)

type Disjoint l r = (WellBehaved l, WellBehaved r, Subset l (l .+ r), Subset r (l .+ r), ((l .+ r) .\\ l) ≈ r, ((l .+ r) .\\ r) ≈ l)

A type synonym for disjointness.

pattern (:+) :: forall l r. Disjoint l r => Rec l -> Rec r -> Rec (l .+ r) infixl 6

A pattern version of record union, for use in pattern matching.

(.//) :: Rec r -> Rec r' -> Rec (r .// r') infixl 6

Record overwrite.

The operation r .// r' creates a new record such that:

Any label that is in both r and r' is in the resulting record with the type and value given by the fields in r,

Any label that is only found in r is in the resulting record.

Any label that is only found in r' is in the resulting record.

This can be thought of as r "overwriting" r'.

Variant construction

pattern IsJust :: forall l r. (AllUniqueLabels r, KnownSymbol l) => Label l -> (r .! l) -> Var r

A pattern for variants; can be used to both destruct a variant when in a pattern position or construct one in an expression position.

Expansion

diversify :: forall r' r. Var r -> Var (r .\/ r')

Make the variant arbitrarily more diverse.

Destruction

impossible :: Var Empty -> a

A Variant with no options is uninhabited.

trial :: KnownSymbol l => Var r -> Label l -> Either (Var (r .- l)) (r .! l)

Convert a variant into either the value at the given label or a variant without that label. This is the basic variant destructor.

trial' :: KnownSymbol l => Var r -> Label l -> Maybe (r .! l)

A version of trial that ignores the leftover variant.

multiTrial :: forall x y. (AllUniqueLabels x, FreeForall x) => Var y -> Either (Var (y .\\ x)) (Var x)

A trial over multiple types

Labels

labels :: forall ρ c s. (IsString s, Forall ρ c) => [s]

Return a list of the labels in a row type.