Data.Row.Internal

Description

This module implements the internals of open records and variants.

Rows

newtype Row a

The kind of rows. This type is only used as a datakind. A row is a typelevel entity telling us which symbols are associated with which types.

Constructors

R [LT a]
A row is a list of symbol-to-type pairs that should always be sorted lexically by the symbol. The constructor is exported here (because this is an internal module) but should not be exported elsewhere.

data Label (s :: Symbol)

A label

Constructors

Label

class KnownSymbol (n :: Symbol)

This class gives the string associated with a type-level symbol. There are instances of the class for every concrete literal: "hello", etc.

Since: base-4.7.0.0

Minimal complete definition

symbolSing

data LT a

The kind of elements of rows. Each element is a label and its associated type.

Constructors

Symbol :-> a

type Empty = R '[]

Type level version of empty

data HideType where

Elements stored in a Row type are usually hidden.

Constructors

HideType :: a -> HideType

Row Operations

type family Extend (l :: Symbol) (a :: k) (r :: Row k) :: Row k where ...

Type level Row extension

Equations

Extend l a (R '[]) = R ((l :-> a) ': '[])

Extend l a (R x) = R (Inject (l :-> a) x)

type family Modify (l :: Symbol) (a :: k) (r :: Row k) :: Row k where ...

Type level Row modification

Equations

Modify l a (R ρ) = R (ModifyR l a ρ)

type family Rename (l :: Symbol) (l' :: Symbol) (r :: Row k) :: Row k where ...

Type level row renaming

Equations

Rename l l' r = Extend l' (r .! l) (r .- l)

type (.==) (l :: Symbol) (a :: k) = Extend l a Empty infix 7

A type level way to create a singleton Row.

type family (r :: Row k) .! (t :: Symbol) :: k where ... infixl 5

Type level label fetching

Equations

(R r) .! l = Get l r

type family (r :: Row k) .- (s :: Symbol) :: Row k where ... infixl 6

Type level Row element removal

Equations

(R r) .- l = R (Remove l r)

type family (l :: Row k) .\\ (r :: Row k) :: Row k where ... infixl 6

Type level Row difference. That is, l .\\ r is the row remaining after removing any matching elements of r from l.

Equations

(R l) .\\ (R r) = R (Diff l r)

Various row-type merges

The difference between .+ (read "append"), .\/ (read "min-join"), and .\\ (read "const-union") comes down to how duplicates are handled. In .+, the two given row-types must be entirely unique. Even the same entry in both row-types is forbidden. In .\/, this final restriction is relaxed, allowing two row-types that have no conflicts to be merged in the logical way. The .\\ operator is the most liberal, allowing any two row-types to be merged together, and whenever there is a conflict, favoring the left argument.

As examples of use:

.+ is used when appending two records, assuring that those two records are entirely disjoint.

.\/ is used when diversifying a variant, allowing some extension to the row-type so long as no original types have changed.

.// is used when doing record overwrite, allowing data in a record to totally overwrite what was previously there.

type family (l :: Row k) .+ (r :: Row k) :: Row k where ... infixl 6

Type level Row append

Equations

x .+ (R '[]) = x

(R '[]) .+ y = y

x .+ (R '[l :-> a]) = Extend l a x

(R '[l :-> a]) .+ y = Extend l a y

(R l) .+ (R r) = R (Merge l r)

type family (l :: Row k) .\/ (r :: Row k) where ... infixl 6

The minimum join of the two rows.

Equations

x .\/ (R '[]) = x

(R '[]) .\/ y = y

(R l) .\/ (R r) = R (MinJoinR l r)

type family (l :: Row k) .// (r :: Row k) where ... infixl 6

The overwriting union, where the left row overwrites the types of the right row where the labels overlap.

Equations

x .// (R '[]) = x

(R '[]) .// y = y

(R l) .// (R r) = R (ConstUnionR l r)

Row Constraints

class Lacks (l :: Symbol) (r :: Row *)

Alias for .\. It is a class rather than an alias, so that it can be partially applied.

type family (r :: Row k) .\ (l :: Symbol) :: Constraint where ... infixl 4

Does the row lack (i.e. it does not have) the specified label?

Equations

(R '[]) .\ l = Unconstrained

(R r) .\ l = LacksR l r r

class (r .! l) ≈ a => HasType l a r

Alias for (r .! l) ≈ a. It is a class rather than an alias, so that it can be partially applied.

class Forall (r :: Row k) (c :: k -> Constraint) where

Any structure over a row in which every element is similarly constrained can be metamorphized into another structure over the same row.

Methods

metamorph

Arguments

:: forall (p :: * -> * -> *) (f :: Row k -> *) (g :: Row k -> *) (h :: k -> *). Bifunctor p

=> Proxy (Proxy h, Proxy p)

-> (f Empty -> g Empty)
The way to transform the empty element

-> (forall ℓ τ ρ. (KnownSymbol ℓ, c τ, HasType ℓ τ ρ) => Label ℓ -> f ρ -> p (f (ρ .- ℓ)) (h τ))
The unfold

-> (forall ℓ τ ρ. (KnownSymbol ℓ, c τ, FrontExtends ℓ τ ρ, AllUniqueLabels (Extend ℓ τ ρ)) => Label ℓ -> p (g ρ) (h τ) -> g (Extend ℓ τ ρ))
The fold

-> f r
The input structure

-> g r

A metamorphism is an anamorphism (an unfold) followed by a catamorphism (a fold). The parameter p describes the output of the unfold and the input of the fold. For records, p = (,), because every entry in the row will unfold to a value paired with the rest of the record. For variants, p = Either, because there will either be a value or future types to explore. Const can be useful when the types in the row are unnecessary.

class BiForall (r1 :: Row k1) (r2 :: Row k2) (c :: k1 -> k2 -> Constraint) where

Any structure over two rows in which the elements of each row satisfy some constraints can be metamorphized into another structure over both of the rows.

Methods

biMetamorph :: forall (p :: * -> * -> *) (f :: Row k1 -> Row k2 -> *) (g :: Row k1 -> Row k2 -> *) (h :: k1 -> k2 -> *). Bifunctor p => Proxy (Proxy h, Proxy p) -> (f Empty Empty -> g Empty Empty) -> (forall ℓ τ1 τ2 ρ1 ρ2. (KnownSymbol ℓ, c τ1 τ2, HasType ℓ τ1 ρ1, HasType ℓ τ2 ρ2) => Label ℓ -> f ρ1 ρ2 -> p (f (ρ1 .- ℓ) (ρ2 .- ℓ)) (h τ1 τ2)) -> (forall ℓ τ1 τ2 ρ1 ρ2. (KnownSymbol ℓ, c τ1 τ2, FrontExtends ℓ τ1 ρ1, FrontExtends ℓ τ2 ρ2, AllUniqueLabels (Extend ℓ τ1 ρ1), AllUniqueLabels (Extend ℓ τ2 ρ2)) => Label ℓ -> p (g ρ1 ρ2) (h τ1 τ2) -> g (Extend ℓ τ1 ρ1) (Extend ℓ τ2 ρ2)) -> f r1 r2 -> g r1 r2

A metamorphism is an anamorphism (an unfold) followed by a catamorphism (a fold).

class (c1 x, c2 y) => BiConstraint c1 c2 x y

A pair of constraints

class Unconstrained

A null constraint

class Unconstrained1 a

A null constraint of one argument

class Unconstrained2 a b

A null constraint of two arguments

class FrontExtends l t r where

A class wrapper for FrontExtendsDict.

Methods

frontExtendsDict :: FrontExtendsDict l t r

data FrontExtendsDict l t r

A dictionary of information that proves that extending a row-type r with a label l will necessarily put it to the front of the underlying row-type list. This is quite internal and should not generally be necessary.

Constructors

forall ρ. FrontExtendsDict (Dict (r ~ R ρ, R ((l :-> t) ': ρ) ≈ Extend l t (R ρ), AllUniqueLabelsR ((l :-> t) ': ρ)))

type WellBehaved ρ = (Forall ρ Unconstrained1, AllUniqueLabels ρ)

A convenient way to provide common, easy constraints

type family AllUniqueLabels (r :: Row k) :: Constraint where ...

Are all of the labels in this Row unique?

Equations

AllUniqueLabels (R r) = AllUniqueLabelsR r

type family Ap (fs :: Row (a -> b)) (r :: Row a) :: Row b where ...

Take two rows with the same labels, and apply the type operator from the first row to the type of the second.

Equations

Ap (R fs) (R r) = R (ApR fs r)

type family ApSingle (fs :: Row (a -> b)) (x :: a) :: Row b where ...

Take a row of type operators and apply each to the second argument.

Equations

ApSingle (R fs) x = R (ApSingleR fs x)

type family Zip (r1 :: Row *) (r2 :: Row *) where ...

Zips two rows together to create a Row of the pairs. The two rows must have the same set of labels.

Equations

Zip (R r1) (R r2) = R (ZipR r1 r2)

type family Map (f :: a -> b) (r :: Row a) :: Row b where ...

Map a type level function over a Row.

Equations

Map f (R r) = R (MapR f r)

type family Subset (r1 :: Row k) (r2 :: Row k) :: Constraint where ...

Is the first row a subset of the second? Or, does the second row contain every binding that the first one does?

Equations

Subset ('R '[]) r = Unconstrained

Subset ('R ((l ':-> a) ': x)) r = ((r .! l) ≈ a, Subset ('R x) r)

type Disjoint l r = (WellBehaved l, WellBehaved r, Subset l (l .+ r), Subset r (l .+ r), ((l .+ r) .\\ l) ≈ r, ((l .+ r) .\\ r) ≈ l)

A type synonym for disjointness.

Helper functions

type family Labels (r :: Row a) where ...

The labels in a Row.

Equations

Labels (R '[]) = '[]

Labels (R ((l :-> a) ': xs)) = l ': Labels (R xs)

labels :: forall ρ c s. (IsString s, Forall ρ c) => [s]

Return a list of the labels in a row type.

labels' :: forall ρ s. (IsString s, Forall ρ Unconstrained1) => [s]

Return a list of the labels in a row type and is specialized to the Unconstrained1 constraint.

show' :: (IsString s, Show a) => a -> s

A helper function for showing labels

toKey :: forall s. KnownSymbol s => Label s -> Text

A helper function to turn a Label directly into Text.

type (≈) a b = a ~ b infix 4

A lower fixity operator for type equality