Data.Row.Variants

Description

This module implements extensible variants using closed type families.

Types and constraints

data Label (s :: Symbol)

A label

Constructors

Label

class KnownSymbol (n :: Symbol)

This class gives the string associated with a type-level symbol. There are instances of the class for every concrete literal: "hello", etc.

Since: base-4.7.0.0

Minimal complete definition

symbolSing

type family AllUniqueLabels (r :: Row k) :: Constraint where ...

Are all of the labels in this Row unique?

Equations

AllUniqueLabels (R r) = AllUniqueLabelsR r

type WellBehaved ρ = (Forall ρ Unconstrained1, AllUniqueLabels ρ)

A convenient way to provide common, easy constraints

data Var (r :: Row *)

The variant type.

data Row a

The kind of rows. This type is only used as a datakind. A row is a typelevel entity telling us which symbols are associated with which types.

type Empty = R '[]

Type level version of empty

type (≈) a b = a ~ b infix 4

A lower fixity operator for type equality

Construction

class (r .! l) ≈ a => HasType l a r

Alias for (r .! l) ≈ a. It is a class rather than an alias, so that it can be partially applied.

pattern IsJust :: forall l r. (AllUniqueLabels r, KnownSymbol l) => Label l -> (r .! l) -> Var r

A pattern for variants; can be used to both destruct a variant when in a pattern position or construct one in an expression position.

singleton :: KnownSymbol l => Label l -> a -> Var (l .== a)

A quick constructor to create a singleton variant.

unSingleton :: forall l a. KnownSymbol l => Var (l .== a) -> (Label l, a)

A quick destructor for singleton variants.

fromLabels :: forall c ρ f. (Alternative f, Forall ρ c, AllUniqueLabels ρ) => (forall l a. (KnownSymbol l, c a) => Label l -> f a) -> f (Var ρ)

Initialize a variant from a producer function that accepts labels. If this function returns more than one possibility, then one is chosen arbitrarily to be the value in the variant.

fromLabelsMap :: forall c f g ρ. (Alternative f, Forall ρ c, AllUniqueLabels ρ) => (forall l a. (KnownSymbol l, c a) => Label l -> f (g a)) -> f (Var (Map g ρ))

Initialize a variant over a Map.

Extension

type family (r :: Row k) .\ (l :: Symbol) :: Constraint where ... infixl 4

Does the row lack (i.e. it does not have) the specified label?

Equations

(R '[]) .\ l = Unconstrained

(R r) .\ l = LacksR l r r

class Lacks (l :: Symbol) (r :: Row *)

Alias for .\. It is a class rather than an alias, so that it can be partially applied.

type family (l :: Row k) .\/ (r :: Row k) where ... infixl 6

The minimum join of the two rows.

Equations

x .\/ (R '[]) = x

(R '[]) .\/ y = y

(R l) .\/ (R r) = R (MinJoinR l r)

diversify :: forall r' r. Var r -> Var (r .\/ r')

Make the variant arbitrarily more diverse.

extend :: forall a l r. KnownSymbol l => Label l -> Var r -> Var (Extend l a r)

A weaker version of diversify, but it's helpful for metamorph as it explicitly uses Extend.

type family (l :: Row k) .+ (r :: Row k) :: Row k where ... infixl 6

Type level Row append

Equations

x .+ (R '[]) = x

(R '[]) .+ y = y

x .+ (R '[l :-> a]) = Extend l a x

(R '[l :-> a]) .+ y = Extend l a y

(R l) .+ (R r) = R (Merge l r)

Modification

update :: (KnownSymbol l, (r .! l) ≈ a) => Label l -> a -> Var r -> Var r

If the variant exists at the given label, update it to the given value. Otherwise, do nothing.

focus :: forall l r r' a b p f. (AllUniqueLabels r, AllUniqueLabels r', KnownSymbol l, (r .! l) ≈ a, (r' .! l) ≈ b, r' ≈ ((r .- l) .\/ (l .== b)), Applicative f, Choice p) => Label l -> p a (f b) -> p (Var r) (f (Var r'))

If the variant exists at the given label, focus on the value associated with it. Otherwise, do nothing.

type family Modify (l :: Symbol) (a :: k) (r :: Row k) :: Row k where ...

Type level Row modification

Equations

Modify l a (R ρ) = R (ModifyR l a ρ)

rename :: (KnownSymbol l, KnownSymbol l') => Label l -> Label l' -> Var r -> Var (Rename l l' r)

Rename the given label.

type family Rename (l :: Symbol) (l' :: Symbol) (r :: Row k) :: Row k where ...

Type level row renaming

Equations

Rename l l' r = Extend l' (r .! l) (r .- l)

Destruction

impossible :: Var Empty -> a

A Variant with no options is uninhabited.

trial :: KnownSymbol l => Var r -> Label l -> Either (Var (r .- l)) (r .! l)

Convert a variant into either the value at the given label or a variant without that label. This is the basic variant destructor.

trial' :: KnownSymbol l => Var r -> Label l -> Maybe (r .! l)

A version of trial that ignores the leftover variant.

multiTrial :: forall x y. (AllUniqueLabels x, FreeForall x) => Var y -> Either (Var (y .\\ x)) (Var x)

A trial over multiple types

view :: KnownSymbol l => Label l -> Var r -> Maybe (r .! l)

A convenient function for using view patterns when dispatching variants. For example:

 myShow :: Var ("y" '::= String :| "x" '::= Int :| Empty) -> String
 myShow (view x -> Just n) = "Int of "++show n
 myShow (view y -> Just s) = "String of "++s

type family Subset (r1 :: Row k) (r2 :: Row k) :: Constraint where ...

Is the first row a subset of the second? Or, does the second row contain every binding that the first one does?

Equations

Subset ('R '[]) r = Unconstrained

Subset ('R ((l ':-> a) ': x)) r = ((r .! l) ≈ a, Subset ('R x) r)

restrict :: forall r r'. (WellBehaved r, Subset r r') => Var r' -> Maybe (Var r)

Arbitrary variant restriction. Turn a variant into a subset of itself.

split :: forall s r. (WellBehaved s, Subset s r) => Var r -> Either (Var (r .\\ s)) (Var s)

Split a variant into two sub-variants.

Types for destruction

type family (r :: Row k) .! (t :: Symbol) :: k where ... infixl 5

Type level label fetching

Equations

(R r) .! l = Get l r

type family (r :: Row k) .- (s :: Symbol) :: Row k where ... infixl 6

Type level Row element removal

Equations

(R r) .- l = R (Remove l r)

type family (l :: Row k) .\\ (r :: Row k) :: Row k where ... infixl 6

Type level Row difference. That is, l .\\ r is the row remaining after removing any matching elements of r from l.

Equations

(R l) .\\ (R r) = R (Diff l r)

type (.==) (l :: Symbol) (a :: k) = Extend l a Empty infix 7

A type level way to create a singleton Row.

Native Conversion

The toNative and fromNative functions allow one to convert between Vars and regular Haskell data types ("native" types) that have the same number of constructors such that each constructor has one field and the same name as one of the options of the Var, which has the same type as that field. As expected, they compose to form the identity. Alternatively, one may use fromNativeGeneral, which allows a variant with excess options to still be transformed to a native type. Because of this, fromNativeGeneral requires a type application (although fromNative does not). The only requirement is that the native Haskell data type be an instance of Generic.

For example, consider the following simple data type:

>>> data Pet = Dog {age :: Int} | Cat {age :: Int} deriving (Generic, Show)

Then, we have the following:

>>> toNative $ IsJust (Label @"Dog") 3 :: Pet
Dog {age = 3}
>>> V.fromNative $ Dog 3 :: Var ("Dog" .== Int .+ "Cat" .== Int)
{Dog=3}

toNative :: ToNative t => Var (NativeRow t) -> t

Convert a variant to a native Haskell type.

fromNative :: FromNative t => t -> Var (NativeRow t)

Convert a Haskell variant to a row-types Var.

fromNativeGeneral :: FromNativeGeneral t ρ => t -> Var ρ

Convert a Haskell variant to a row-types Var.

type ToNative t = (Generic t, ToNativeG (Rep t))

type FromNative t = (Generic t, FromNativeG (Rep t))

type FromNativeGeneral t ρ = (Generic t, FromNativeGeneralG (Rep t) ρ)

type family NativeRow t where ...

Equations

NativeRow t = NativeRowG (Rep t)

Row operations

Map

type family Map (f :: a -> b) (r :: Row a) :: Row b where ...

Map a type level function over a Row.

Equations

Map f (R r) = R (MapR f r)

map :: forall c f r. Forall r c => (forall a. c a => a -> f a) -> Var r -> Var (Map f r)

A function to map over a variant given a constraint.

map' :: forall f r. FreeForall r => (forall a. a -> f a) -> Var r -> Var (Map f r)

A function to map over a variant given no constraint.

transform :: forall c r f g. Forall r c => (forall a. c a => f a -> g a) -> Var (Map f r) -> Var (Map g r)

Lifts a natrual transformation over a variant. In other words, it acts as a variant transformer to convert a variant of f a values to a variant of g a values. If no constraint is needed, instantiate the first type argument with Unconstrained1.

transform' :: forall r f g. FreeForall r => (forall a. f a -> g a) -> Var (Map f r) -> Var (Map g r)

A form of transformC that doesn't have a constraint on a

Fold

class Forall (r :: Row k) (c :: k -> Constraint)

Any structure over a row in which every element is similarly constrained can be metamorphized into another structure over the same row.

Minimal complete definition

metamorph

erase :: forall c ρ b. Forall ρ c => (forall a. c a => a -> b) -> Var ρ -> b

A standard fold

eraseWithLabels :: forall c ρ s b. (Forall ρ c, IsString s) => (forall a. c a => a -> b) -> Var ρ -> (s, b)

A fold with labels

eraseZipGeneral :: forall c ρ b s. (Forall ρ c, IsString s) => (forall x y. (c x, c y) => Either (s, x, x) ((s, x), (s, y)) -> b) -> Var ρ -> Var ρ -> b

A fold over two variants at once. A call eraseZipGeneral f x y will return f (Left (show l, a, b)) when x and y both have values at the same label l and will return f (Right ((show l1, a), (show l2, b))) when they have values at different labels l1 and l2 respectively.

eraseZip :: forall c ρ b. Forall ρ c => (forall a. c a => a -> a -> b) -> Var ρ -> Var ρ -> Maybe b

A simpler fold over two variants at once

Applicative-like functions

traverse :: forall c f r. (Forall r c, Functor f) => (forall a. c a => a -> f a) -> Var r -> f (Var r)

Traverse a function over a variant.

traverseMap :: forall c f g h r. (Forall r c, Functor f) => (forall a. c a => g a -> f (h a)) -> Var (Map g r) -> f (Var (Map h r))

Traverse a function over a Mapped variant.

sequence :: forall f r. (FreeForall r, Functor f) => Var (Map f r) -> f (Var r)

Applicative sequencing over a variant

Compose

We can easily convert between mapping two functors over the types of a row and mapping the composition of the two functors. The following two functions perform this composition with the gaurantee that:

>>> compose . uncompose = id

>>> uncompose . compose = id

compose :: forall f g r. FreeForall r => Var (Map f (Map g r)) -> Var (Map (Compose f g) r)

Convert from a variant where two functors have been mapped over the types to one where the composition of the two functors is mapped over the types.

uncompose :: forall f g r. FreeForall r => Var (Map (Compose f g) r) -> Var (Map f (Map g r))

Convert from a variant where the composition of two functors have been mapped over the types to one where the two functors are mapped individually one at a time over the types.

labels

labels :: forall ρ c s. (IsString s, Forall ρ c) => [s]

Return a list of the labels in a row type.

ApSingle functions

eraseSingle :: forall c fs x y. Forall fs c => (forall f. c f => f x -> y) -> Var (ApSingle fs x) -> y

A version of erase that works even when the row-type of the variant argument is of the form ApSingle fs x.

mapSingle :: forall c fs x y. Forall fs c => (forall f. c f => f x -> f y) -> Var (ApSingle fs x) -> Var (ApSingle fs y)

Performs a functorial-like map over an ApSingle variant. In other words, it acts as a variant transformer to convert a variant of f x values to a variant of f y values. If no constraint is needed, instantiate the first type argument with Unconstrained1.

mapSingleA :: forall c fs g x y. (Forall fs c, Functor g) => (forall f. c f => f x -> g (f y)) -> Var (ApSingle fs x) -> g (Var (ApSingle fs y))

Like mapSingle, but works over a functor.

eraseZipSingle :: forall c fs x y z. Forall fs c => (forall f. c f => f x -> f y -> z) -> Var (ApSingle fs x) -> Var (ApSingle fs y) -> Maybe z

A version of eraseZip that works even when the row-types of the variant arguments are of the form ApSingle fs x.

Coerce

coerceVar :: forall r1 r2. BiForall r1 r2 Coercible => Var r1 -> Var r2

Coerce a variant to a coercible representation. The BiForall in the context indicates that the type of any option in r1 can be coerced to the type of the corresponding option in r2.

Internally, this is implemented just with unsafeCoerce, but we provide the following implementation as a proof:

newtype ConstV a b = ConstV { unConstV :: Var a }
newtype ConstV a b = FlipConstV { unFlipConstV :: Var b }
coerceVar :: forall r1 r2. BiForall r1 r2 Coercible => Var r1 -> Var r2
coerceVar = unFlipConstV . biMetamorph @_ @_ @r1 @r2 @Coercible @Either @ConstV @FlipConstV @Const Proxy doNil doUncons doCons . ConstV
  where
    doNil = impossible . unConstV
    doUncons l = bimap ConstV Const . flip trial l . unConstV
    doCons :: forall ℓ τ1 τ2 ρ1 ρ2. (KnownSymbol ℓ, Coercible τ1 τ2, AllUniqueLabels (Extend ℓ τ2 ρ2))
           => Label ℓ -> Either (FlipConstV ρ1 ρ2) (Const τ1 τ2)
           -> FlipConstV (Extend ℓ τ1 ρ1) (Extend ℓ τ2 ρ2)
    doCons l (Left (FlipConstV v)) = FlipConstV $ extend @τ2 l v
    doCons l (Right (Const x)) = FlipConstV $ IsJust l (coerce @τ1 @τ2 x)
      \\ extendHas @ρ2 @ℓ @τ2